renamed DC tools; ready to test Docker Compose on Win+WSL2
