+++ firebase-dc-tools/.dockerignore
