+++ firebase-dc-tools/firebase-prime/.gitignore
