+++ firebase-dc-tools/firebase-prime/.npmrc
