+++ firebase-dc-tools/firebase-prime/src/config.js
