+++ firebase-dc-tools/firebase-prime/src/index.js
