+++ firebase-dc-tools/firebase-prime/src/primeData/wipe.js
